refactor: 요청, 응답에서 쓰이는 객체 common 패키지로 이동
