+++ tomcat/src/main/java/org/apache/catalina/controller/LoginController.java
@@ -5,8 +5,8 @@ import static org.reflections.Reflections.log;
 import org.apache.catalina.session.Session;
 import org.apache.catalina.session.SessionIdGenerator;
 import org.apache.catalina.session.SessionManager;
-import org.apache.coyote.http11.Headers;
-import org.apache.coyote.http11.HttpCookies;
+import org.apache.coyote.http11.common.Headers;
+import org.apache.coyote.http11.common.HttpCookies;
 import org.apache.coyote.http11.request.Request;
 import org.apache.coyote.http11.request.RequestBody;
 import org.apache.coyote.http11.request.RequestLine;
